% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CxF.R
\name{CxF}
\alias{CxF}
\title{Title CxF Columnas a Filas
Sort a dataframe that contains "responses in multiple columns" in a single column}
\usage{
CxF(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
a CSV file
}
\description{
Title CxF Columnas a Filas
Sort a dataframe that contains "responses in multiple columns" in a single column
}
