% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrizRelacional.R
\name{matrizRelacional}
\alias{matrizRelacional}
\title{Title matrizRelacional}
\usage{
matrizRelacional(data_df, criterio1, criterio2)
}
\arguments{
\item{data_df}{dataframe that will be worked}

\item{criterio1}{criterion of the column "Questions"}

\item{criterio2}{criterion of the column "Questions"}
}
\value{
An relational matrix
}
\description{
Function that transforms a data set into a relationship matrix
between two criteria. The required format of the input dataframe is 7 columns:
 "Id", "Time.of.Screening", "Title", "Authors", "Question", "List.of.Reviewers"
  and "Answers".
}
