% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{sankey}
\alias{sankey}
\title{Title sankey}
\usage{
sankey(matriz, title = "Basic Sankey Diagram")
}
\arguments{
\item{matriz}{adjacency matrix}

\item{title}{plot title}
}
\value{
A sankey diagram plot
}
\description{
PLot a sankey diagram from a matrix
}
\examples{
sankey(matrix(40, 4, 10))
}
